Improved exception handling

+++ src/main/java/ru/mipt/messenger/services/UserService.java
@@ -1,7 +1,11 @@
 package ru.mipt.messenger.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import ru.mipt.messenger.exceptions.ResourceNotFoundException;
 import ru.mipt.messenger.models.User;
 import ru.mipt.messenger.repositories.UserRepository;
 
@@ -16,23 +20,78 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public void createUser(User user) {
+    /**
+     * Creates a new user in the database. Please make sure that request body matches the following criteria:
+     * 1) userId is null. This is required to correctly generate userId in the database.
+     * It is not required to explicitly assign userId to null in the JSON since it defaults to null.
+     * 2) nickname, firstname, phone, email and password are not null, not blank, have valid length and match patterns
+     * if required.
+     * 3) nickname, firstname, phone and email are unique in the database.
+     * 4) registrationDttm is not null and matches pattern "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
+     * @see <a href="https://docs.spring.io/spring-framework/docs/3.2.4.RELEASE_to_4.0.0.M3/Spring%20Framework%203.2.4.RELEASE/org/springframework/format/annotation/DateTimeFormat.ISO.html">Spring docs</a>
+     * 5) dateOfBirth is not null and matches pattern "yyyy-MM-dd".
+     * 6) If secondName is not null, it has valid length.
+     * 7) If status is not null, it is either "Banned" or "Active". If status is null, it defaults to "Active".
+     * 8) If role is not null, it is either "Admin" or "User". If role is null. it defaults to "User".
+     * 9) If profilePictureLink is not null, it has valid length.
+     * @throws DataIntegrityViolationException if userId is not null or other database error occurs.
+     * @throws MethodArgumentNotValidException if any field of user is invalid.
+     * @throws HttpMessageNotReadableException if JSON is invalid.
+     * @param user User to create
+     */
+    public void createUser(User user) throws DataIntegrityViolationException, HttpMessageNotReadableException {
+        if (user.getUserId() != null) {
+            throw new DataIntegrityViolationException("UserID must be null for new database entries");
+        }
+
         userRepository.save(user);
     }
 
+    /**
+     * Reads user by userId.
+     * @param id userId to read.
+     * @return if such userId is present in the database, returns the user. If there is no user with such userId,
+     * returns null.
+     */
     public User readUserById(Integer id) {
-        return userRepository.findById(id).get();
+        var result = userRepository.findById(id);
+
+        return result.orElse(null);
     }
 
-    public List<User> readUsersByNickname(String nickname) {
-        return userRepository.findUsersByNickname(nickname);
+    /**
+     * Reads users by firstname.
+     * @param firstname firstname to read.
+     * @return list of users whose firstname equals required one. If there are no such users, returns an empty list.
+     */
+    public List<User> readUsersByFirstname(String firstname) {
+        return userRepository.findUsersByFirstname(firstname);
     }
 
-    public void updateUser(Integer id, User newUser) {
-        userRepository.save(newUser);
+    /**
+     * Updates an existing user. Please note that you have to send the entire JSON in request body, even if you want
+     * to update a single field.
+     * @param updateData user to update.
+     * @throws ResourceNotFoundException if user with updateData.userId does not exist.
+     */
+    public void updateUser(User updateData) throws ResourceNotFoundException {
+        if (!userRepository.existsById(updateData.getUserId())) {
+            throw new ResourceNotFoundException("User with userId " + updateData.getUserId() + " does not exist");
+        }
+
+        userRepository.save(updateData);
     }
 
-    public void deleteUser(Integer id) {
+    /**
+     * Deletes an existing user.
+     * @param id user to delete.
+     * @throws ResourceNotFoundException if user with updateData.userId does not exist.
+     */
+    public void deleteUser(Integer id) throws ResourceNotFoundException {
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User with userId " + id + " does not exist");
+        }
+
         userRepository.deleteById(id);
     }
 }
